// ux/api/client/types.ts
export interface components {
    schemas: {
// ...
        TermRelationshipCreate: {
            /**
             * Source Term Id
             * Format: uuid
             */
            source_term_id: string;
            /**
             * Target Term Id
             * Format: uuid
             */
            target_term_id: string;
            /** Predicate */
            predicate: string;
        };
        /** TermRelationshipOut */
        TermRelationshipOut: {
            /**
             * Source Term Id
             * Format: uuid
             */
            source_term_id: string;
            /**
             * Target Term Id
             * Format: uuid
             */
            target_term_id: string;
            /** Predicate */
            predicate: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Created At */
            created_at: string;
            /** Source Term */
            source_term?: {
                [key: string]: unknown;
            } | null;
            /** Target Term */
            target_term?: {
                [key: string]: unknown;
            } | null;
        };
        /** TermRelationshipUpdate */
        TermRelationshipUpdate: {
            /** Predicate */
            predicate: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}